requirements.txt
discord.py>=2.3.0
discord-ext-voice-recv>=0.5.0
google-generativeai>=0.7.0
google-cloud-speech>=2.26.0
python-dotenv>=1.0.0
PyNaCl>=1.5.0
gTTS>=2.5.0
aiohttp>=3.9.0
